:white_check_mark: simplification of direct CO2
Direct CO2 accounts for the stoichiometric emissions due to the combustion of a resource. It has been removed from the RESSOURCE table has it wasn't use. It has been relovated in the documentation. The values are used to assess the CO2 emissions for the usage of these resources through different technologies.